# apis/openapi.yaml
openapi: 3.0.3
info:
  title: Facility Layout API
  version: "1.0.0"
  description: >
    The system of record for **where things physically are in the building**,
    for the warehouse-systems platform. This bounded context owns a site's
    structural hierarchy — Site → Area → Zone → Aisle → Bay → Level →
    Position — and the coded storage slots inside it. It is the "warehouse
    map" that every other context reads and none of them writes.


    Its aggregates are **Site** (a physical facility, the root of the
    hierarchy), **Zone** (a behavioral classification scoped to a Site,
    bundling the Area and Zone code segments, carrying a TemperatureClass and
    a Hazmat flag), **Aisle** (a physical corridor scoped to a Zone, carrying
    the SequenceHint that gives downstream travel-path planning its
    walk-order input), **LocationType** + **PlacementRule** (which kinds of
    slot are legal in which zones), and **LocationSlot** (the coded leaf
    location, whose identity IS its LocationCode).


    A LocationCode is a value object, never free text: seven typed segments,
    hyphen-joined coarsest to finest, each non-empty and `[A-Z0-9]` only —
    for example `WH1-STOR-AMB-A07-03-02-B`. Registering a slot is a
    chain-of-custody check, not a bare insert: the Site/Zone/Aisle chain the
    code implies must resolve to existing, Active aggregates, and the slot's
    LocationType must satisfy every PlacementRule applicable to its Zone.
...
    every caller.


...


...
  contact:
...
servers:
...
security: []
...
paths:
...
  /sites:
    post:
      operationId: registerSite
      summary: Register a physical facility
      description: >
        Adds a Site to the warehouse map. The site code becomes the first
        segment of every LocationCode inside the facility, so it must be
        uppercase alphanumeric and unique across the service. Site code
        uniqueness is enforced here in the application layer rather than in
        the aggregate: a single aggregate cannot see its siblings. A newly
        registered Site is Active, and publishes SiteRegistered.
      tags: [Sites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSiteRequest"
            example:
              siteCode: WH1
              name: Fulfilment Centre One
      responses:
        "201":
          description: Site registered.
          headers:
            Location:
              description: Path of the newly created Site.
              schema:
                type: string
                example: /sites/WH1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
              example:
                siteCode: WH1
                name: Fulfilment Centre One
                status: Active
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listSites
      summary: List every registered site
      description: >
        Returns every Site on the warehouse map, ordered by site code,
        including sites that have been decommissioned (their status says so).
        This is a read model over the Site repository; it performs no writes
        and publishes no events.
      tags: [Sites]
      responses:
        "200":
          description: Every registered site, ordered by code.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
              example:
                - siteCode: WH1
                  name: Fulfilment Centre One
                  status: Active
                - siteCode: WH2
                  name: Fulfilment Centre Two
                  status: Decommissioned
        "500":
          $ref: "#/components/responses/InternalError"
  /sites/{siteCode}:
    get:
      operationId: getSite
      summary: Get one site
      description: >
        Returns a single Site by its code. This is the URL the Location
        header of a successful registerSite points at.
      tags: [Sites]
      parameters:
        - $ref: "#/components/parameters/SiteCode"
      responses:
        "200":
          description: The site.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
              example:
                siteCode: WH1
                name: Fulfilment Centre One
                status: Active
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sites/{siteCode}/zones:
    post:
      operationId: registerZone
      summary: Register a behavioral zone inside a site's area
      description: >
        Adds a Zone scoped to the given Site. The zone's identity is
        `SITE-AREA-ZONE` (e.g. `WH1-STOR-AMB`), which is exactly the first
        three segments of every LocationCode inside it. A Zone cannot be
        registered against an unknown or non-Active Site. The TemperatureClass
        and Hazmat flag set here are what PlacementRules match on later, so
        they are behavioral, not descriptive. Publishes ZoneRegistered.
      tags: [Zones]
      parameters:
        - $ref: "#/components/parameters/SiteCode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterZoneRequest"
            example:
              areaCode: STOR
              zoneCode: FRZ
              temperatureClass: Frozen
              hazmat: false
      responses:
        "201":
          description: Zone registered.
          headers:
            Location:
              description: Path of the newly created Zone.
              schema:
                type: string
                example: /zones/WH1-STOR-FRZ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
              example:
                zoneId: WH1-STOR-FRZ
                siteCode: WH1
                areaCode: STOR
                zoneCode: FRZ
                temperatureClass: Frozen
                hazmat: false
                status: Active
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listZones
      summary: List a site's zones
      description: >
        Returns every Zone in the given Site, ordered by zone id. Answers 404
        if the site itself is unknown, rather than an empty list, so a caller
        can tell "no zones yet" apart from "no such site".
      tags: [Zones]
      parameters:
        - $ref: "#/components/parameters/SiteCode"
      responses:
        "200":
          description: The site's zones, ordered by zone id.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
              example:
                - zoneId: WH1-RCV-AMB
                  siteCode: WH1
                  areaCode: RCV
                  zoneCode: AMB
                  temperatureClass: Ambient
                  hazmat: false
                  status: Active
                - zoneId: WH1-STOR-AMB
                  siteCode: WH1
                  areaCode: STOR
                  zoneCode: AMB
                  temperatureClass: Ambient
                  hazmat: false
                  status: Active
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sites/{siteCode}/layout:
    get:
      operationId: getSiteLayout
      summary: Get a site's full drawable layout
      description: >
        The primary "draw the warehouse" endpoint. Returns the whole
        structure of one Site as a single nested document — zones → aisles →
        location slots — with every slot's exploded coordinates, type,
        capacity and status. Zones come back ordered by id, aisles in
        SequenceHint **walk order** (not registration order), and slots
        ordered bay → level → position, so a frontend can render a floor plan
        directly with no client-side joining or sorting.


        With `?format=svg` the identical read model is rendered server-side
        as a minimal SVG floor plan (one coloured band per zone, one row of
        rects per aisle, colour = temperature class / hazmat) that a human
        can curl straight into a `.svg` file and open in a browser.


        This is a PROJECTION assembled by querying across the aggregates, not
        separately stored state. It performs no writes and publishes no
        events.
      tags: [Layout]
      parameters:
        - $ref: "#/components/parameters/SiteCode"
        - name: format
          in: query
          required: false
          description: >
            Response representation. Omit (or use `json`) for the nested JSON
            document; use `svg` to receive an `image/svg+xml` floor plan of
            the same data.
          schema:
            type: string
            enum: [json, svg]
            default: json
          example: svg
      responses:
        "200":
          description: >
            The site's full layout, as nested JSON or as an SVG floor plan
            depending on the `format` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteLayout"
              example:
                site:
                  siteCode: WH1
                  name: Fulfilment Centre One
                  status: Active
                zones:
                  - zoneId: WH1-STOR-AMB
                    siteCode: WH1
                    areaCode: STOR
                    zoneCode: AMB
                    temperatureClass: Ambient
                    hazmat: false
                    status: Active
                    aisles:
                      - aisleId: WH1-STOR-AMB-A07
                        zoneId: WH1-STOR-AMB
                        aisleCode: A07
                        sequenceHint: 7
                        direction: TwoWay
                        status: Active
                        slots:
                          - locationCode: WH1-STOR-AMB-A07-03-02-B
                            zoneId: WH1-STOR-AMB
                            aisleId: WH1-STOR-AMB-A07
                            coordinates:
                              site: WH1
                              area: STOR
                              zone: AMB
                              aisle: A07
                              bay: "03"
                              level: "02"
                              position: B
                            locationType: PalletRack
                            capacity:
                              maxWeightKg: 1200
                              maxVolumeM3: 2.4
                            status: Active
                totals:
                  zones: 1
                  aisles: 1
                  slots: 1
            image/svg+xml:
              schema:
                type: string
                description: A standalone SVG document.
              example: |
                <svg xmlns="http://www.w3.org/2000/svg" width="640" height="240" viewBox="0 0 640 240">
                  <title>WH1 facility layout</title>
                </svg>
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /zones/{zoneId}:
    get:
      operationId: getZone
      summary: Get one zone
      description: >
        Returns a single Zone by its id (`SITE-AREA-ZONE`). This is the URL
        the Location header of a successful registerZone points at.
      tags: [Zones]
      parameters:
        - $ref: "#/components/parameters/ZoneId"
      responses:
        "200":
          description: The zone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
              example:
                zoneId: WH1-STOR-AMB
                siteCode: WH1
                areaCode: STOR
                zoneCode: AMB
                temperatureClass: Ambient
                hazmat: false
                status: Active
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /zones/{zoneId}/grid:
    get:
      operationId: getZoneGrid
      summary: Get a zone as a 2D renderable grid
      description: >
        The second "draw the warehouse" endpoint. Returns one Zone's slots as
        an explicit matrix: rows are Levels, columns are (Aisle, Bay) pairs
        in aisle SequenceHint walk order, and `rows[i].cells` is
        index-aligned with `columns`. A cell is literal JSON `null` wherever
        the rack has a gap; otherwise it lists the Positions at that
        coordinate. A UI iterates rows × columns and paints — no client-side
        layout maths.


        Like the site layout this is a pure projection: no writes, no events.
      tags: [Layout]
      parameters:
        - $ref: "#/components/parameters/ZoneId"
      responses:
        "200":
          description: The zone's slots as a level × (aisle, bay) matrix.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneGrid"
              example:
                zone:
                  zoneId: WH1-STOR-AMB
                  siteCode: WH1
                  areaCode: STOR
                  zoneCode: AMB
                  temperatureClass: Ambient
                  hazmat: false
                  status: Active
                columns:
                  - aisleId: WH1-STOR-AMB-A07
                    aisleCode: A07
                    bay: "03"
                    sequenceHint: 7
                  - aisleId: WH1-STOR-AMB-A09
                    aisleCode: A09
                    bay: "01"
                    sequenceHint: 9
                levels: ["01", "02"]
                rows:
                  - level: "01"
                    cells:
                      - positions:
                          - locationCode: WH1-STOR-AMB-A07-03-01-A
                            position: A
                            locationType: PalletRack
                            status: Active
                      - positions:
                          - locationCode: WH1-STOR-AMB-A09-01-01-A
                            position: A
                            locationType: Shelf
                            status: Active
                  - level: "02"
                    cells:
                      - positions:
                          - locationCode: WH1-STOR-AMB-A07-03-02-A
                            position: A
                            locationType: PalletRack
                            status: Active
                          - locationCode: WH1-STOR-AMB-A07-03-02-B
                            position: B
                            locationType: PalletRack
                            status: Active
                      - null
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /zones/{zoneId}/aisles:
    post:
      operationId: registerAisle
      summary: Register a physical corridor inside a zone
      description: >
        Adds an Aisle scoped to the given Zone. Its identity is the zone id
        plus the aisle code (e.g. `WH1-STOR-AMB-A07`), the first four
        segments of every LocationCode inside it. The SequenceHint is the
        aisle's walk-order position — lower comes earlier on the route — and
        is the concrete travel-distance input downstream WES planning
        consumes. An Aisle cannot be registered against an unknown or
        non-Active Zone. Publishes AisleRegistered.
      tags: [Aisles]
      parameters:
        - $ref: "#/components/parameters/ZoneId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAisleRequest"
            example:
              aisleCode: A07
              sequenceHint: 7
              direction: TwoWay
      responses:
        "201":
          description: Aisle registered.
          headers:
            Location:
              description: Path of the newly created Aisle.
              schema:
                type: string
                example: /zones/WH1-STOR-AMB/aisles/A07
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aisle"
              example:
                aisleId: WH1-STOR-AMB-A07
                zoneId: WH1-STOR-AMB
                aisleCode: A07
                sequenceHint: 7
                direction: TwoWay
                status: Active
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listAisles
      summary: List a zone's aisles in walk order
      description: >
        Returns every Aisle in the given Zone, ordered by SequenceHint then
        aisle code — that is, in the order an associate would walk them.
        Answers 404 if the zone itself is unknown.
      tags: [Aisles]
      parameters:
        - $ref: "#/components/parameters/ZoneId"
      responses:
        "200":
          description: The zone's aisles, in walk order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aisle"
              example:
                - aisleId: WH1-STOR-AMB-A07
                  zoneId: WH1-STOR-AMB
                  aisleCode: A07
                  sequenceHint: 7
                  direction: TwoWay
                  status: Active
                - aisleId: WH1-STOR-AMB-A09
                  zoneId: WH1-STOR-AMB
                  aisleCode: A09
                  sequenceHint: 9
                  direction: OneWay
                  status: Active
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /zones/{zoneId}/aisles/{aisleCode}:
    get:
      operationId: getAisle
      summary: Get one aisle
      description: >
        Returns a single Aisle by its zone id and aisle code. This is the URL
        the Location header of a successful registerAisle points at.
      tags: [Aisles]
      parameters:
        - $ref: "#/components/parameters/ZoneId"
        - name: aisleCode
          in: path
          required: true
          description: The aisle segment of a LocationCode.
          schema:
            type: string
            pattern: "^[A-Z0-9]+$"
          example: A07
      responses:
        "200":
          description: The aisle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aisle"
              example:
                aisleId: WH1-STOR-AMB-A07
                zoneId: WH1-STOR-AMB
                aisleCode: A07
                sequenceHint: 7
                direction: TwoWay
                status: Active
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /location-types/{name}:
    get:
      operationId: getLocationType
      summary: Get one location type
      description: >
        Returns a single LocationType by name. This is the URL the Location
        header of a successful registerLocationType points at.
      tags: [Location Types]
      parameters:
        - name: name
          in: path
          required: true
          description: The location type's name.
          schema:
            type: string
          example: PalletRack
      responses:
        "200":
          description: The location type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationType"
              example:
                name: PalletRack
                defaultCapacity:
                  maxWeightKg: 1200
                  maxVolumeM3: 2.4
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
...
    RegisterSiteRequest:
      type: object
      required: [siteCode, name]
      properties:
        siteCode:
          type: string
          pattern: "^[A-Z0-9]+$"
          description: Uppercase alphanumeric code. Becomes the first LocationCode segment.
          example: WH1
        name:
          type: string
          minLength: 1
          description: Human-readable facility name.
          example: Fulfilment Centre One
...